Barriers.jl

Barrier factories

Barriers.Barrier — Type

Barrier(ntasks::Integer) -> barrier

Creates a barrier for ntasks tasks. Call cycle!(barrier[i]) in i-th task for waiting for other tasks to arrive at the same phase.

Barriers.fuzzy_barrier — Function

fuzzy_barrier(ntasks::Integer) -> barrier::Barrier

Missing docstring.

Missing docstring for fuzzy_reduce_barrier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reduce_barrier. Check Documenter's build log for details.

Barrier constructors

Missing docstring.

Missing docstring for CentralizedBarrier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DisseminationBarrier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FlatTreeBarrier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FuzzyBarrier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StaticTreeBarrier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TreeBarrier. Check Documenter's build log for details.

Synchronizing operations

Barriers.cycle! — Function

cycle!(barrier[i])

Using a barrier::Barrier, signal that i::Integer-th thread reached a certain phase of the program and wait for other tasks to reach the same phase.

Examples

julia> using Barriers

julia> xs = [1:3;];

julia> barrier = Barrier(3);

julia> @sync for i in 1:3
           Threads.@spawn begin
               x = i^2
               xs[i] = x
               cycle!(barrier[i])
               xs[mod1(i + 1, 3)] -= x
           end
       end

julia> xs
3-element Vector{Int64}:
 -8
  3
  5

Barriers.arrive! — Function

arrive!(barrier[i])

Signal that i::Integer-th thread reached a certain phase but postpone the synchronization for the departure.

Note that not all Barrier subtypes support arrive!.

See fuzzy_barrier, depart!.

Barriers.depart! — Function

depart!(barrier[i])

Wait for all calls to arrive!(barrier[i]) or reduce_arrive!(barrier[i], _) for i = 1, 2, ..., ntasks.

Note that not all Barrier subtypes support depart!.

See fuzzy_barrier, arrive!, reduce_arrive!.

Missing docstring.

Missing docstring for reduce!. Check Documenter's build log for details.

Barriers.reduce_arrive! — Function

reduce_arrive!(barrier[i])